requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
tqdm==4.66.1
# 以下包仅在数据迁移时需要
#pymysql==1.1.0 